+++ CHANGELOG.md
@@ -2,8 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.0] - 2025-06-25
+## [1.0.0] - 2025-06-30
 
 ### Collection
 
 - Initial release.
+- Added *ilo_account_settings* module plugin.
+- Added *ilo_domain_name* module plugin.
+- Added *ilo_hostname* module plugin.
+- Added *ilo_https_certificate* module plugin.
+- Added *ilo_https_certificate_info* module plugin.
+- Added *ilo_https_csr* module plugin.
+- Added *ilo_https_security_settings* module plugin.
+- Added *ilo_ipv4_dns_servers* module plugin.
+- Added *ilo_ipv6_dns_servers* module plugin.
+- Added *ilo_ldap_remote_role_mapping* module plugin.
+- Added *ilo_ldap_settings* module plugin.
+- Added *ilo_ntp_servers* module plugin.
+- Added *ilo_snmp_community* module plugin.
+- Added *ilo_snmp_config* module plugin.
+- Added *ilo_snmp_user* module plugin.
+- Added *ilo_time_zone* module plugin.
+- Added *ilo_https_certificate* role.